+++ library/src/common/action/actions/verify_can_write.rs
@@ -1,4 +1,4 @@
-build_action! { VerifyCanWrite (src_prog, hook, arg_id, args, _act_args, do_return, return_value) {
+build_action! { VerifyCanWrite (par_prog, src_prog, hook, arg_id, args, _act_args, do_return, return_value) {
         let directory_index = args.iter().position(|arg| arg.id == arg_id).expect("WhiteBeam: Lost track of environment");
         let directory_argument: crate::common::db::ArgumentRow = args[directory_index].clone();
         let library: &str = &hook.library;
@@ -46,7 +46,7 @@ build_action! { VerifyCanWrite (src_prog, hook, arg_id, args, _act_args, do_retu
         let class = String::from("Filesystem/Directory/Writable");
         let all_allowed_directories: Vec<String> = {
             let whitelist_cache_lock = crate::common::db::WL_CACHE.lock().expect("WhiteBeam: Failed to lock mutex");
-            whitelist_cache_lock.iter().filter(|whitelist| (whitelist.class == class) && ((whitelist.path == src_prog) || (whitelist.path == any))).map(|whitelist| whitelist.value.clone()).collect()
+            whitelist_cache_lock.iter().filter(|whitelist| (whitelist.class == class) && ((whitelist.parent == par_prog) || (whitelist.parent == any)) && ((whitelist.path == src_prog) || (whitelist.path == any))).map(|whitelist| whitelist.value.clone()).collect()
         };
         // Permit ANY
         if all_allowed_directories.iter().any(|directory| directory == &any) {