+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Application-specific whitelists
 - Modular test framework (compile time reflection)
+- Parent process whitelisting
 
 ### Changed